[GEOG-10] feat: Add weekly ranking service & persistence for weekly rankings

+++ src/main/java/com/pretz/geographic/application/domain/model/WeeklyPosition.java
@@ -2,7 +2,7 @@ package com.pretz.geographic.application.domain.model;
 
 import java.util.Objects;
 
-public record WeeklyPosition(Game game, Week week, Player player, int wins, int points) {
+public record WeeklyPosition(Game game, Week week, Player player, int wins, float points) {
 
     public WeeklyPosition {
         Objects.requireNonNull(game, "WeeklyPosition game must not be null");
